+++ client/src/graphql/generated.ts
@@ -245,6 +245,9 @@ export type GQLAddPoliciesResponse = {
 };
 
 export type GQLAddPolicyInput = {
+  readonly applyUserStrikeCountConfigToChildren?: InputMaybe<
+    Scalars['Boolean']['input']
+  >;
   readonly enforcementGuidelines?: InputMaybe<Scalars['String']['input']>;
   readonly id?: InputMaybe<Scalars['ID']['input']>;
   readonly name: Scalars['String']['input'];
@@ -252,6 +255,7 @@ export type GQLAddPolicyInput = {
   readonly parentName?: InputMaybe<Scalars['String']['input']>;
   readonly policyText?: InputMaybe<Scalars['String']['input']>;
   readonly policyType?: InputMaybe<GQLPolicyType>;
+  readonly userStrikeCount?: InputMaybe<Scalars['Int']['input']>;
 };
 
 export type GQLAggregation = {
@@ -387,6 +391,16 @@ export type GQLBaseField = GQLField & {
   readonly name: Scalars['String']['output'];
   readonly required: Scalars['Boolean']['output'];
   readonly type: GQLFieldType;
+};
+
+export type GQLBuiltInActionImmutableError = GQLError & {
+  readonly __typename: 'BuiltInActionImmutableError';
+  readonly detail?: Maybe<Scalars['String']['output']>;
+  readonly pointer?: Maybe<Scalars['String']['output']>;
+  readonly requestId?: Maybe<Scalars['String']['output']>;
+  readonly status: Scalars['Int']['output'];
+  readonly title: Scalars['String']['output'];
+  readonly type: ReadonlyArray<Scalars['String']['output']>;
 };
 
 export type GQLCannotDeleteDefaultUserError = GQLError & {
@@ -1507,6 +1521,16 @@ export type GQLIntegrationNoInputCredentialsError = GQLError & {
   readonly type: ReadonlyArray<Scalars['String']['output']>;
 };
 
+export type GQLInvalidActionItemTypeIdsError = GQLError & {
+  readonly __typename: 'InvalidActionItemTypeIdsError';
+  readonly detail?: Maybe<Scalars['String']['output']>;
+  readonly pointer?: Maybe<Scalars['String']['output']>;
+  readonly requestId?: Maybe<Scalars['String']['output']>;
+  readonly status: Scalars['Int']['output'];
+  readonly title: Scalars['String']['output'];
+  readonly type: ReadonlyArray<Scalars['String']['output']>;
+};
+
 export type GQLInvalidItemTypeHiddenFieldsError = GQLError & {
   readonly __typename: 'InvalidItemTypeHiddenFieldsError';
   readonly detail?: Maybe<Scalars['String']['output']>;
@@ -1519,6 +1543,16 @@ export type GQLInvalidItemTypeHiddenFieldsError = GQLError & {
 
 export type GQLInvalidItemTypeSchemaError = GQLError & {
   readonly __typename: 'InvalidItemTypeSchemaError';
+  readonly detail?: Maybe<Scalars['String']['output']>;
+  readonly pointer?: Maybe<Scalars['String']['output']>;
+  readonly requestId?: Maybe<Scalars['String']['output']>;
+  readonly status: Scalars['Int']['output'];
+  readonly title: Scalars['String']['output'];
+  readonly type: ReadonlyArray<Scalars['String']['output']>;
+};
+
+export type GQLInvalidPolicyParentError = GQLError & {
+  readonly __typename: 'InvalidPolicyParentError';
   readonly detail?: Maybe<Scalars['String']['output']>;
   readonly pointer?: Maybe<Scalars['String']['output']>;
   readonly requestId?: Maybe<Scalars['String']['output']>;
@@ -2432,7 +2466,10 @@ export const GQLMutateActionError = {
 export type GQLMutateActionError =
   (typeof GQLMutateActionError)[keyof typeof GQLMutateActionError];
 export type GQLMutateActionResponse =
-  GQLActionNameExistsError | GQLMutateActionSuccessResponse;
+  | GQLActionNameExistsError
+  | GQLBuiltInActionImmutableError
+  | GQLInvalidActionItemTypeIdsError
+  | GQLMutateActionSuccessResponse;
 
 export type GQLMutateActionSuccessResponse = {
   readonly __typename: 'MutateActionSuccessResponse';
@@ -3474,6 +3511,16 @@ export type GQLPolicyActionCount = {
   readonly count: Scalars['Int']['output'];
   readonly itemSubmissionIds: ReadonlyArray<Scalars['String']['output']>;
   readonly policyId?: Maybe<Scalars['String']['output']>;
+};
+
+export type GQLPolicyHierarchyCycleError = GQLError & {
+  readonly __typename: 'PolicyHierarchyCycleError';
+  readonly detail?: Maybe<Scalars['String']['output']>;
+  readonly pointer?: Maybe<Scalars['String']['output']>;
+  readonly requestId?: Maybe<Scalars['String']['output']>;
+  readonly status: Scalars['Int']['output'];
+  readonly title: Scalars['String']['output'];
+  readonly type: ReadonlyArray<Scalars['String']['output']>;
 };
 
 export type GQLPolicyNameExistsError = GQLError & {
@@ -4943,7 +4990,11 @@ export type GQLUpdatePolicyInput = {
   readonly userStrikeCount?: InputMaybe<Scalars['Int']['input']>;
 };
 
-export type GQLUpdatePolicyResponse = GQLNotFoundError | GQLPolicy;
+export type GQLUpdatePolicyResponse =
+  | GQLInvalidPolicyParentError
+  | GQLNotFoundError
+  | GQLPolicy
+  | GQLPolicyHierarchyCycleError;
 
 export type GQLUpdateReportingRuleInput = {
   readonly actionIds?: InputMaybe<ReadonlyArray<Scalars['ID']['input']>>;
@@ -5906,6 +5957,8 @@ export type GQLCreateActionMutation = {
         readonly status: number;
         readonly type: ReadonlyArray<string>;
       }
+    | { readonly __typename: 'BuiltInActionImmutableError' }
+    | { readonly __typename: 'InvalidActionItemTypeIdsError' }
     | {
         readonly __typename: 'MutateActionSuccessResponse';
         readonly data: {
@@ -5955,6 +6008,8 @@ export type GQLUpdateActionMutation = {
         readonly status: number;
         readonly type: ReadonlyArray<string>;
       }
+    | { readonly __typename: 'BuiltInActionImmutableError' }
+    | { readonly __typename: 'InvalidActionItemTypeIdsError' }
     | {
         readonly __typename: 'MutateActionSuccessResponse';
         readonly data: {
@@ -19450,6 +19505,7 @@ export type GQLUpdatePolicyMutationVariables = Exact<{
 export type GQLUpdatePolicyMutation = {
   readonly __typename: 'Mutation';
   readonly updatePolicy:
+    | { readonly __typename: 'InvalidPolicyParentError' }
     | { readonly __typename: 'NotFoundError' }
     | {
         readonly __typename: 'Policy';
@@ -19461,7 +19517,8 @@ export type GQLUpdatePolicyMutation = {
         readonly policyType?: GQLPolicyType | null;
         readonly userStrikeCount?: number | null;
         readonly applyUserStrikeCountConfigToChildren?: boolean | null;
-      };
+      }
+    | { readonly __typename: 'PolicyHierarchyCycleError' };
 };
 
 export type GQLDeletePolicyMutationVariables = Exact<{