+++ server/graphql/modules/action.ts
@@ -204,9 +204,29 @@ const typeDefs = /* GraphQL */ `
     requestId: String
   }
 
+  type InvalidActionItemTypeIdsError implements Error {
+    title: String!
+    status: Int!
+    type: [String!]!
+    pointer: String
+    detail: String
+    requestId: String
+  }
+
+  type BuiltInActionImmutableError implements Error {
+    title: String!
+    status: Int!
+    type: [String!]!
+    pointer: String
+    detail: String
+    requestId: String
+  }
+
   union MutateActionResponse =
     | MutateActionSuccessResponse
     | ActionNameExistsError
+    | InvalidActionItemTypeIdsError
+    | BuiltInActionImmutableError
 
   type MutateActionSuccessResponse {
     data: CustomAction!
@@ -439,6 +459,12 @@ const Mutation: GQLMutationResolvers = {
       if (isCoopErrorOfType(e, 'ActionNameExistsError')) {
         return gqlErrorResult(e, `/input/name`);
       }
+      if (isCoopErrorOfType(e, 'InvalidActionItemTypeIdsError')) {
+        return gqlErrorResult(e, `/input/itemTypeIds`);
+      }
+      if (isCoopErrorOfType(e, 'BuiltInActionImmutableError')) {
+        return gqlErrorResult(e);
+      }
 
       throw e;
     }
@@ -458,6 +484,12 @@ const Mutation: GQLMutationResolvers = {
     } catch (e: unknown) {
       if (isCoopErrorOfType(e, 'ActionNameExistsError')) {
         return gqlErrorResult(e, `/input/name`);
+      }
+      if (isCoopErrorOfType(e, 'InvalidActionItemTypeIdsError')) {
+        return gqlErrorResult(e, `/input/itemTypeIds`);
+      }
+      if (isCoopErrorOfType(e, 'BuiltInActionImmutableError')) {
+        return gqlErrorResult(e);
       }
 
       throw e;